[seller] refactor(215): constant 파일 하나로 통일
