--- apps/seller/src/features/products/create/create-form-options/options-shipping-days.constant.ts
@@ -1,9 +1,0 @@
-export const SHIPPING_DAYS = [
-  { label: '월', value: 'mon' },
-  { label: '화', value: 'tue' },
-  { label: '수', value: 'wed' },
-  { label: '목', value: 'thu' },
-  { label: '금', value: 'fri' },
-  { label: '토', value: 'sat' },
-  { label: '일', value: 'sun' },
-]
